Update README.md
This Admin Dashboard is a practice project that I built to learn and experiment with React, TypeScript, and Chakra UI. The application provides an admin interface to manage dashboard settings, profile details, and products, and it showcases a CRUD table for product management.

As a beginner project, my goal was to understand how to integrate TypeScript for type safety and improve my frontend skills by utilizing Chakra UI, a popular and modular component library for building modern web applications. I also implemented private routing to ensure that sensitive parts of the app are accessible only to authenticated users.

This project was built primarily for learning purposes and is a simple yet comprehensive example of using modern React features like routing, state management, and interacting with an external API. The backend functionality is simulated using MockAPI, which allows CRUD operations on the product data.